+++ src/mwax_mover/mwa_archiver.py
@@ -16,7 +16,7 @@ import random
 import time
 import uuid
 import logging
-
+from mwax_mover.utils import running_under_pytest
 from mwax_mover.mwax_command import run_command_ext
 
 logger = logging.getLogger(__name__)
@@ -179,12 +179,14 @@ def archive_file_rclone(
     full_filename: str,
     bucket_name: str,
     md5hash: str,
-):
+) -> bool:
     """Upload a file to Pawsey S3 (Acacia/Banksia) via rclone with retry.
 
     Attempts upload to random endpoints from the provided list with checksum
     verification. Retries on failure with remaining endpoints until all are
     exhausted.
+
+    if running under pytest, will automatically return True
 
     Args:
         rclone_profile: The rclone profile/remote name to use.
@@ -204,6 +206,7 @@ def archive_file_rclone(
     # get file size
     try:
         file_size = os.path.getsize(full_filename)
+        size_gigabytes = float(file_size) / (1000.0 * 1000.0 * 1000.0)
     except Exception:
         logger.exception(f"{full_filename}: Error determining file size.")
         return False
@@ -225,40 +228,55 @@ def archive_file_rclone(
         #  test.txt banksia:/mwaingest-14322
         #
         try:
-            cmdline = f'/usr/bin/rclone copyto -M --metadata-set "md5={md5hash}" --s3-endpoint={endpoint_url} {full_filename} {rclone_profile}:/{bucket_name}/{filename}'
+            #
+            # TODO: Ugly solution here for testing- should replace this with a Mock pattern
+            #
+            if running_under_pytest():
+                # Return true
+                elapsed = 1.0
+                gbps_per_sec = (size_gigabytes * 8.0) / elapsed
+                check_elapsed = 1.0
 
-            # run rclone copyto
-            return_val, stdout = run_command_ext(cmdline, None, 600, False)
-
-            if return_val:
-                elapsed = time.time() - start_time
-                size_gigabytes = float(file_size) / (1000.0 * 1000.0 * 1000.0)
-                gbps_per_sec = (size_gigabytes * 8) / elapsed
-
-                # Success - now verify the file at the remote
-                logger.debug(
-                    f"{full_filename}: attempting check against {rclone_profile} {endpoint_url} bucket {bucket_name}..."
+                logger.info(
+                    f"{full_filename}: archive_file_rclone success."
+                    f"Copied ({size_gigabytes:.3f}GB in {elapsed:.3f} seconds at"
+                    f" {gbps_per_sec:.3f} Gbps). Check took {check_elapsed:.3f} seconds."
                 )
-                cmdline = f"/usr/bin/rclone check --s3-endpoint={endpoint_url} {full_filename} {rclone_profile}:/{bucket_name}"
+                return True
+            else:
+                cmdline = f'/usr/bin/rclone copyto -M --metadata-set "md5={md5hash}" --s3-endpoint={endpoint_url} {full_filename} {rclone_profile}:/{bucket_name}/{filename}'
 
-                # run rclone check
+                # run rclone copyto
                 return_val, stdout = run_command_ext(cmdline, None, 600, False)
 
                 if return_val:
-                    # If checksums match then rclone returns exit code 0. Otherwise !=0.
-                    # run_command_ext returns True for 0 and False for anything else
-                    check_elapsed = time.time() - start_time
+                    elapsed = time.time() - start_time
+                    gbps_per_sec = (size_gigabytes * 8) / elapsed
 
-                    logger.info(
-                        f"{full_filename}: archive_file_rclone success."
-                        f"Copied ({size_gigabytes:.3f}GB in {elapsed:.3f} seconds at"
-                        f" {gbps_per_sec:.3f} Gbps). Check took {check_elapsed:.3f} seconds."
+                    # Success - now verify the file at the remote
+                    logger.debug(
+                        f"{full_filename}: attempting check against {rclone_profile} {endpoint_url} bucket {bucket_name}..."
                     )
-                    return True
+                    cmdline = f"/usr/bin/rclone check --s3-endpoint={endpoint_url} {full_filename} {rclone_profile}:/{bucket_name}"
+
+                    # run rclone check
+                    return_val, stdout = run_command_ext(cmdline, None, 600, False)
+
+                    if return_val:
+                        # If checksums match then rclone returns exit code 0. Otherwise !=0.
+                        # run_command_ext returns True for 0 and False for anything else
+                        check_elapsed = time.time() - start_time
+
+                        logger.info(
+                            f"{full_filename}: archive_file_rclone success."
+                            f"Copied ({size_gigabytes:.3f}GB in {elapsed:.3f} seconds at"
+                            f" {gbps_per_sec:.3f} Gbps). Check took {check_elapsed:.3f} seconds."
+                        )
+                        return True
+                    else:
+                        raise Exception(stdout)
                 else:
                     raise Exception(stdout)
-            else:
-                raise Exception(stdout)
         except Exception:
             logger.exception(
                 f"{full_filename}: Error uploading to {endpoint_url} bucket {bucket_name} via rclone."