+++ tests/test001_subdis_beamformer_subfile.py
@@ -10,6 +10,7 @@ import time
 
 from mwax_mover.utils import MWAXSubfileDistirbutorMode
 from mwax_mover.mwax_subfile_distributor import MWAXSubfileDistributor
+from tests_fakedb import FakeMWAXDBHandler
 from tests_common import setup_test_directories, create_observation_subfiles
 
 TEST_CONFIG_FILE = "tests/data/test001/test001.cfg"
@@ -27,7 +28,13 @@ def test_beamformer_subfile():
 
     # Create a subfile distributor
     sd = MWAXSubfileDistributor()
-    sd.initialise(TEST_CONFIG_FILE, MWAXSubfileDistirbutorMode.BEAMFORMER)
+
+    # Override db_handler with a fake one
+    fake_db_handler = FakeMWAXDBHandler()
+    # Add any select results (in order in the code below-or keep commented if none)
+    # e.g. fake_db_handler.select_results = [[{"observation_num": 123, "size": 1024, "checksum": "abc123"}]]
+
+    sd.initialise(TEST_CONFIG_FILE, MWAXSubfileDistirbutorMode.BEAMFORMER, fake_db_handler)
 
     # setup data
     metafits = os.path.join(sd.cfg_corr_metafits_path, os.path.basename(TEST_METAFITS))