+++ tests/test016_calvin_controller.py
@@ -1,26 +1,25 @@
 import logging
 import os
-from mwax_mover.mwax_calvin_processor import MWAXCalvinProcessor, CalvinJobType
+from mwax_mover.mwax_calvin_controller import MWAXCalvinController
 import tests_common
 from tests_fakedb import FakeMWAXDBHandler
 
 logger = logging.getLogger(__name__)
 
 
-def test_mwax_calvin_processor():
+def test_mwax_calvin_controller():
     """Tests that mwax_calvin reads a config file ok"""
     # Setup all the paths
-    base_dir = tests_common.setup_test_directories("test013")
+    base_dir = tests_common.setup_test_directories("test016")
 
     # Start mwax_subfile_distributor using our test config
-    mcal = MWAXCalvinProcessor()
-    mcal.produce_debug_plots = False
+    mcal = MWAXCalvinController()
 
     # Override the hostname
     mcal.hostname = "calvin99"
 
     # Determine config file location
-    config_filename = "tests/data/test013/test013.cfg"
+    config_filename = "tests/data/test016/test016.cfg"
 
     # Override db_handler with a fake one
     fake_db_handler = FakeMWAXDBHandler()
@@ -30,13 +29,6 @@ def test_mwax_calvin_processor():
     # Call to read config <-- this is what we're testing!
     mcal.initialise(
         config_filename,
-        1234567890,
-        1234,
-        CalvinJobType.realtime,
-        "https//test.test",
-        [
-            123,
-        ],
         fake_db_handler,
     )
 
@@ -51,16 +43,5 @@ def test_mwax_calvin_processor():
 
     assert mcal.health_multicast_interface_name == "eth0"
     assert mcal.health_multicast_ip == "127.0.0.1"
-    assert mcal.health_multicast_port == 8011
+    assert mcal.health_multicast_port == 8012
     assert mcal.health_multicast_hops == 1
-
-    assert mcal.job_input_path == os.path.join(base_dir, "data/calvin/in_jobs")
-    assert mcal.job_output_path == os.path.join(base_dir, "data/calvin/out_jobs")
-    assert mcal.source_list_filename == "../srclists/srclist_pumav3_EoR0aegean_fixedEoR1pietro+ForA_phase1+2.txt"
-    assert mcal.source_list_type == "rts"
-    assert mcal.hyperdrive_binary_path == "../mwa_hyperdrive/target/release/hyperdrive"
-    assert mcal.hyperdrive_timeout == 7200
-    assert mcal.birli_binary_path == "../Birli/target/release/birli"
-    assert mcal.birli_timeout == 3600
-    assert mcal.keep_completed_visibility_files == 0
-    assert mcal.cal_export_max_age_hours == 24